finish first test

+++ birra-meet-service/src/main/java/com/santander/birrameet/domain/Location.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @Getter
 public class Location {
 
-    private Integer longitude;
-    private Integer latitude;
+    private Double longitude;
+    private Double latitude;
 }
